🎨: デザインシステム連携のため削除ダイアログパターンの記述構成を見直す
